refactor: split config e2e scenarios into focused files
Reorganize config-related e2e tests by user-facing behavior so file names describe each contract, and replace implementation-oriented wording in style e2e comments with specification-oriented descriptions.
