+++ tests/e2e/config/mod.rs
@@ -1,1 +1,3 @@
-mod config;
+mod config_path;
+mod edit_command;
+mod token_customization;
